Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -5,8 +5,8 @@ char* welcome() {
 }
 
 int main() {
-    int radius;
-    scanf("%d",&radius);
+    float radius;
+    scanf("%f",&radius);
     printf("Area: %f",3.14*radius*radius);
     return 0;
 }
